Fix RedMica::VERSION autoloading for Zeitwerk compatibility
