+++ lib/redmine/version.rb
@@ -44,25 +44,3 @@ module Redmine
     def self.to_s; STRING end
   end
 end
-
-module RedMica
-  module VERSION
-    MAJOR = 3
-    MINOR = 1
-    TINY  = 2
-
-    BRANCH = 'devel'
-
-    # Retrieves the revision from the working copy
-    def self.revision
-      nil
-    end
-
-    REVISION = self.revision
-    ARRAY    = [MAJOR, MINOR, TINY, BRANCH, REVISION].compact
-    STRING   = ARRAY.join('.')
-
-    def self.to_a; ARRAY  end
-    def self.to_s; STRING end
-  end
-end
